customers-service/pyproject.toml
[tool.poetry]
name = "gRPC FastAPI Project"
version = "0.0.1"
description = "Customers Service FastAPI"
authors = ["EscapeFromHell"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = ">=0.110.1"
uvicorn = ">=0.18.3"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
httpx = ">=0.23.0"
requests = "^2.28.1"
urllib3 = "1.26.16"
xlrd = "^2.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
boto3 = "^1.36.26"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"

black = ">=22.1.0"
isort = ">=5.9.3"

flake8 = ">=3.9.2"
flake8-broken-line = ">=0.3.0"
flake8-bugbear = ">=21.9.1"
flake8-builtins = ">=1.5.3"
flake8-coding = ">=1.3.2"
flake8-comprehensions = ">=3.6.1"
flake8-debugger = ">=4.0.0"
flake8-eradicate = ">=1.1.0"
flake8-logging-format = ">=0.6.0"
flake8-pep3101 = ">=1.3.0"
flake8-print = ">=4.0.0"
flake8-pytest-style = ">=1.5.0"
flake8-string-format = ">=0.3.0"
flake8-variables-names = ">=0.0.4"
flake8-expression-complexity = ">=0.0.9"
flake8-cognitive-complexity = ">=0.1.0"
flake8-annotations = ">=2.6.2"
flake8-noqa = ">=1.1.0"
flake8-docstrings = ">=1.6.0"
flake8-unused-arguments = ">=0.0.6"
pep8-naming = ">=0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.git         # exclude a few common directories in the
    | \.pytest_cache          # root of the project
    | \.idea
    | \__pycache__
    | \.venv
    | alembic
    | env
    | venv
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 110
multi_line_output = 3
include_trailing_comma = true
skip = [".git", ".pytest_cache", ".idea", "__pycache__", ".venv", "alembic", "env", "venv"]
